coordinates from double to int

+++ Coordinates.h
@@ -5,18 +5,18 @@ class Boss; class Hero; class Destructible; class Projectile; class Game;
 
 class Coordinates {
 private:
-    double c_x;
-    double c_y;
+    int c_x;
+    int c_y;
     int* g_map;
 
 public:
-    Coordinates(double x, double y, int* map);
-    double getX() { return c_x; };
-    double getY() { return c_y; };
+    Coordinates(int x, int y, int* map);
+    int getX() { return c_x; };
+    int getY() { return c_y; };
     int* getMap() { return g_map; };
     void setX(double x);
     void setY(double y);
-    bool contact(Coordinates* c, double hitBox);
+    bool contact(Coordinates* c, int hitBox);
     bool move(int direction, int id);
     bool murred(int d);
 };
