+++ Projectile.h
@@ -11,10 +11,10 @@ private:
     int p_damages;
     bool p_players; //evitons le suicide et de pouvoir nous faire mal avec nos propres projectiles
     int p_direction; //le tir se déplace vers un point cardinal {N, NE, E, SE, S, SW, W, NW}
-    double p_hitbox;
+    int p_hitbox;
 
 public:
-    Projectile(bool players, int direction, double size, int damages, double x, double y, int* map);
+    Projectile(bool players, int direction, int size, int damages, int x, int y, int* map);
     Coordinates* getCoordinates() { return &p_position; }
     int damages() { return p_damages; }
     bool playerProjectile() { return p_players; }
